Cast to uint16_t and make hex value lowercase

+++ src/credentials/GroupDataProvider.h
@@ -257,7 +257,10 @@ public:
     GroupDataProvider(const GroupDataProvider &)             = delete;
     GroupDataProvider & operator=(const GroupDataProvider &) = delete;
 
-    uint16_t GetMaxGroupsPerFabric() { return IsGroupcastEnabled() ? (getMaxMembershipCount() / 2) : mMaxGroupsPerFabric; }
+    uint16_t GetMaxGroupsPerFabric()
+    {
+        return static_cast<uint16_t>(IsGroupcastEnabled() ? (getMaxMembershipCount() / 2) : mMaxGroupsPerFabric);
+    }
     uint16_t GetMaxGroupKeysPerFabric() const { return mMaxGroupKeysPerFabric; }
 
     /**
